requirements.txt
streamlit

# PDF reading
pypdf

# LangChain core
langchain
langchain-community
langchain-text-splitters

# Embeddings
sentence-transformers
torch

# Vector DB
qdrant-client

# API + Utilities
requests
tqdm
python-dotenv

# EURI (only if used)
euriai
